refactor: Use PlaceholderAvatar everywhere it is needed
