+++ src/script/components/Avatar/AvatarBackground.tsx
