+++ src/script/components/Avatar/AvatarBadge.tsx
