+++ src/script/components/Avatar/AvatarBorder.tsx
