+++ src/script/components/Avatar/AvatarImage.tsx
@@ -28,6 +28,7 @@ import {CSS_FILL_PARENT} from 'Util/CSSMixin';
 
 import {AssetRemoteData} from '../../assets/AssetRemoteData';
 import {AssetRepository} from '../../assets/AssetRepository';
+
 import {AVATAR_SIZE} from '.';
 
 export interface AvatarImageProps {
