+++ src/script/components/Avatar/GroupAvatar.test.tsx
