+++ src/script/components/Avatar/GroupAvatar.tsx
