+++ src/script/components/Avatar/PlaceholderAvatar/DefaultAvatarImage.tsx
