+++ src/script/components/Avatar/PlaceholderAvatar/index.ts
