+++ src/script/components/Avatar/ServiceAvatar.tsx
@@ -30,6 +30,7 @@ import {AvatarImage} from './AvatarImage';
 import {AvatarWrapper} from './AvatarWrapper';
 
 import {ServiceEntity} from '../../integration/ServiceEntity';
+
 import {AVATAR_SIZE} from '.';
 
 export interface ServiceAvatarProps extends React.HTMLProps<HTMLDivElement> {
