+++ src/script/components/Avatar/TemporaryGuest.test.tsx
@@ -22,6 +22,7 @@ import {render} from '@testing-library/react';
 import {TemporaryGuestAvatar} from './TemporaryGuestAvatar';
 
 import {User} from '../../entity/User';
+
 import {AVATAR_SIZE, STATE} from '.';
 jest.mock('../../auth/util/SVGProvider');
