+++ src/script/components/Avatar/TemporaryGuestAvatar.tsx
@@ -29,6 +29,7 @@ import {AvatarWrapper} from './AvatarWrapper';
 import {shouldShowBadge} from './UserAvatar';
 
 import {User} from '../../entity/User';
+
 import {AVATAR_SIZE, STATE, DIAMETER} from '.';
 
 export interface TemporaryGuestAvatarProps extends React.HTMLProps<HTMLDivElement> {
