+++ src/script/components/Avatar/UserAvatar.test.tsx
@@ -22,6 +22,7 @@ import {render} from '@testing-library/react';
 import {UserAvatar} from './UserAvatar';
 
 import {User} from '../../entity/User';
+
 import {AVATAR_SIZE, STATE} from '.';
 jest.mock('../../auth/util/SVGProvider');
