+++ src/script/components/Avatar/index.ts
