+++ src/script/components/MessagesList/Message/ContentMessage/MessageHeader.tsx
@@ -17,8 +17,7 @@
  *
  */
 
-import {AVATAR_SIZE, Avatar} from 'Components/Avatar-m';
-import {PlaceholderAvatar} from 'Components/Avatar-m/PlaceholderAvatar';
+import {AVATAR_SIZE, Avatar} from 'Components/Avatar';
 import {Icon} from 'Components/Icon';
 import {ContentMessage} from 'src/script/entity/message/ContentMessage';
 import {DeleteMessage} from 'src/script/entity/message/DeleteMessage';
@@ -96,16 +95,12 @@ export function MessageHeader({
   return (
     <div className="message-header">
       <div className="message-header-icon">
-        {!isAvailable ? (
-          <PlaceholderAvatar onClick={() => onClickAvatar(sender)} size={AVATAR_SIZE.X_SMALL} />
-        ) : (
-          <Avatar
-            tabIndex={focusTabIndex}
-            participant={sender}
-            onAvatarClick={onClickAvatar}
-            avatarSize={AVATAR_SIZE.X_SMALL}
-          />
-        )}
+        <Avatar
+          tabIndex={focusTabIndex}
+          participant={sender}
+          onAvatarClick={onClickAvatar}
+          avatarSize={AVATAR_SIZE.X_SMALL}
+        />
       </div>
 
       <div className="message-header-label" data-uie-name={uieName}>
