+++ src/script/components/list/ConversationListCell.tsx
@@ -32,8 +32,8 @@ import cx from 'classnames';
 import {Availability} from '@wireapp/protocol-messaging';
 
 import {AvailabilityState} from 'Components/AvailabilityState';
-import {Avatar, AVATAR_SIZE} from 'Components/Avatar-m';
-import {GroupAvatar} from 'Components/Avatar-m/GroupAvatar';
+import {Avatar, AVATAR_SIZE} from 'Components/Avatar';
+import {GroupAvatar} from 'Components/Avatar/GroupAvatar';
 import {Icon} from 'Components/Icon';
 import {useKoSubscribableChildren} from 'Util/ComponentUtil';
 import {isKey, isOneOfKeys, KEY} from 'Util/KeyboardUtil';
