+++ src/script/components/panel/ServiceDetails.tsx
@@ -19,7 +19,7 @@
 
 import React from 'react';
 
-import {Avatar, AVATAR_SIZE} from 'Components/Avatar-m';
+import {Avatar, AVATAR_SIZE} from 'Components/Avatar';
 import {useKoSubscribableChildren} from 'Util/ComponentUtil';
 
 import type {ServiceEntity} from '../../integration/ServiceEntity';
